Implement plug-in based loading of widgets.
- Fix preview widget running refresh upon start, now updates upon show event.
- Refactor printing of JSON file to use log.debug()
- Separated accordion widget into accordion.py
- Moved resources (icons) into their own folder
- Simplified storage of plug-in widgets in app into one attribute (App.plugins)
