+++ capture_gui/app.py
@@ -1,632 +1,17 @@
-import inspect
 import json
 import logging
 import os
 import tempfile
 
 import capture
+import maya.cmds as cmds
 
 from .vendor.Qt import QtCore, QtWidgets, QtGui
-import maya.cmds as cmds
 from . import lib
-from . import widgets
+from . import plugin
+from .accordion import AccordionWidget
 
 log = logging.getLogger("Capture Gui")
-
-
-# Modification of Blur's Accordion Widget to include a Maya style.
-
-class AccordionItem(QtGui.QGroupBox):
-    trigger = QtCore.Signal(bool)
-
-    def __init__(self, accordion, title, widget):
-        QtGui.QGroupBox.__init__(self, accordion)
-
-        # create the layout
-        layout = QtGui.QVBoxLayout()
-        layout.setContentsMargins(6, 12, 6, 6)
-        layout.setSpacing(0)
-        layout.addWidget(widget)
-
-        self._accordianWidget = accordion
-        self._rolloutStyle = 2
-        self._dragDropMode = 0
-
-        self.setAcceptDrops(True)
-        self.setLayout(layout)
-        self.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-        self.customContextMenuRequested.connect(self.showMenu)
-
-        # create custom properties
-        self._widget = widget
-        self._collapsed = False
-        self._collapsible = True
-        self._clicked = False
-        self._customData = {}
-
-        # set common properties
-        self.setTitle(title)
-
-    def accordionWidget(self):
-        """
-            \remarks	grabs the parent item for the accordian widget
-            \return		<blurdev.gui.widgets.accordianwidget.AccordianWidget>
-        """
-        return self._accordianWidget
-
-    def customData(self, key, default=None):
-        """
-            \remarks	return a custom pointer to information stored with this item
-            \param		key			<str>
-            \param		default		<variant>	default value to return if the key was not found
-            \return		<variant> data
-        """
-        return self._customData.get(str(key), default)
-
-    def dragEnterEvent(self, event):
-        if not self._dragDropMode:
-            return
-
-        source = event.source()
-        if source != self and source.parent() == self.parent() and isinstance(
-                source, AccordionItem):
-            event.acceptProposedAction()
-
-    def dragDropRect(self):
-        return QtCore.QRect(25, 7, 10, 6)
-
-    def dragDropMode(self):
-        return self._dragDropMode
-
-    def dragMoveEvent(self, event):
-        if not self._dragDropMode:
-            return
-
-        source = event.source()
-        if source != self and source.parent() == self.parent() and isinstance(
-                source, AccordionItem):
-            event.acceptProposedAction()
-
-    def dropEvent(self, event):
-        widget = event.source()
-        layout = self.parent().layout()
-        layout.insertWidget(layout.indexOf(self), widget)
-        self._accordianWidget.emitItemsReordered()
-
-    def expandCollapseRect(self):
-        return QtCore.QRect(0, 0, self.width(), 20)
-
-    def enterEvent(self, event):
-        self.accordionWidget().leaveEvent(event)
-        event.accept()
-
-    def leaveEvent(self, event):
-        self.accordionWidget().enterEvent(event)
-        event.accept()
-
-    def mouseReleaseEvent(self, event):
-        if self._clicked and self.expandCollapseRect().contains(event.pos()):
-            self.toggleCollapsed()
-            event.accept()
-        else:
-            event.ignore()
-
-        self._clicked = False
-
-    def mouseMoveEvent(self, event):
-        event.ignore()
-
-    def mousePressEvent(self, event):
-        # handle an internal move
-
-        # start a drag event
-        if event.button() == QtCore.Qt.LeftButton and self.dragDropRect().contains(
-                event.pos()):
-            # create the pixmap
-            pixmap = QtGui.QPixmap.grabWidget(self, self.rect())
-
-            # create the mimedata
-            mimeData = QtCore.QMimeData()
-            mimeData.setText('ItemTitle::%s' % (self.title()))
-
-            # create the drag
-            drag = QtGui.QDrag(self)
-            drag.setMimeData(mimeData)
-            drag.setPixmap(pixmap)
-            drag.setHotSpot(event.pos())
-
-            if not drag.exec_():
-                self._accordianWidget.emitItemDragFailed(self)
-
-            event.accept()
-
-        # determine if the expand/collapse should occur
-        elif event.button() == QtCore.Qt.LeftButton and self.expandCollapseRect().contains(
-                event.pos()):
-            self._clicked = True
-            event.accept()
-
-        else:
-            event.ignore()
-
-    def isCollapsed(self):
-        return self._collapsed
-
-    def isCollapsible(self):
-        return self._collapsible
-
-    def __drawTriangle(self, painter, x, y):
-
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255, 160),
-                             QtCore.Qt.SolidPattern)
-        if not self.isCollapsed():
-            tl, tr, tp = QtCore.QPoint(x + 9, y + 8), QtCore.QPoint(x + 19,
-                                                                    y + 8), QtCore.QPoint(
-                x + 14, y + 13.0)
-            points = [tl, tr, tp]
-            triangle = QtGui.QPolygon(points)
-        else:
-            tl, tr, tp = QtCore.QPoint(x + 11, y + 6), QtCore.QPoint(x + 16,
-                                                                     y + 11), QtCore.QPoint(
-                x + 11, y + 16.0)
-            points = [tl, tr, tp]
-            triangle = QtGui.QPolygon(points)
-
-        currentBrush = painter.brush()
-        painter.setBrush(brush)
-        painter.drawPolygon(triangle)
-        painter.setBrush(currentBrush)
-
-    def paintEvent(self, event):
-        painter = QtGui.QPainter()
-        painter.begin(self)
-        painter.setRenderHint(painter.Antialiasing)
-        font = painter.font()
-        font.setBold(True)
-        painter.setFont(font)
-
-        x = self.rect().x()
-        y = self.rect().y()
-        w = self.rect().width() - 1
-        h = self.rect().height() - 1
-        r = 8
-
-        # draw a rounded style
-        if self._rolloutStyle == 2:
-            # draw the text
-            painter.drawText(x + 33, y + 3, w, 16,
-                             QtCore.Qt.AlignLeft | QtCore.Qt.AlignTop,
-                             self.title())
-
-            # draw the triangle
-            self.__drawTriangle(painter, x, y)
-
-            # draw the borders
-            pen = QtGui.QPen(self.palette().color(QtGui.QPalette.Light))
-            pen.setWidthF(0.6)
-            painter.setPen(pen)
-
-            painter.drawRoundedRect(x + 1, y + 1, w - 1, h - 1, r, r)
-
-            pen.setColor(self.palette().color(QtGui.QPalette.Shadow))
-            painter.setPen(pen)
-
-            painter.drawRoundedRect(x, y, w - 1, h - 1, r, r)
-
-        # draw a square style
-        if self._rolloutStyle == 3:
-            # draw the text
-            painter.drawText(x + 33, y + 3, w, 16,
-                             QtCore.Qt.AlignLeft | QtCore.Qt.AlignTop,
-                             self.title())
-
-            self.__drawTriangle(painter, x, y)
-
-            # draw the borders
-            pen = QtGui.QPen(self.palette().color(QtGui.QPalette.Light))
-            pen.setWidthF(0.6)
-            painter.setPen(pen)
-
-            painter.drawRect(x + 1, y + 1, w - 1, h - 1)
-
-            pen.setColor(self.palette().color(QtGui.QPalette.Shadow))
-            painter.setPen(pen)
-
-            painter.drawRect(x, y, w - 1, h - 1)
-
-        # draw a Maya style
-        if self._rolloutStyle == 4:
-            # draw the text
-            painter.drawText(x + 33, y + 3, w, 16,
-                             QtCore.Qt.AlignLeft | QtCore.Qt.AlignTop,
-                             self.title())
-
-            painter.setRenderHint(QtGui.QPainter.Antialiasing, False)
-
-            self.__drawTriangle(painter, x, y)
-
-            # draw the borders - top
-            headerHeight = 20
-
-            headerRect = QtCore.QRect(x + 1, y + 1, w - 1, headerHeight)
-            headerRectShadow = QtCore.QRect(x - 1, y - 1, w + 1,
-                                            headerHeight + 2)
-
-            # Highlight
-            pen = QtGui.QPen(self.palette().color(QtGui.QPalette.Light))
-            pen.setWidthF(0.4)
-            painter.setPen(pen)
-
-            painter.drawRect(headerRect)
-            painter.fillRect(headerRect, QtGui.QColor(255, 255, 255, 18))
-
-            # Shadow
-            pen.setColor(self.palette().color(QtGui.QPalette.Dark))
-            painter.setPen(pen)
-            painter.drawRect(headerRectShadow)
-
-            if not self.isCollapsed():
-                # draw the lover border
-                pen = QtGui.QPen(self.palette().color(QtGui.QPalette.Dark))
-                pen.setWidthF(0.8)
-                painter.setPen(pen)
-
-                offSet = headerHeight + 3
-                bodyRect = QtCore.QRect(x, y + offSet, w, h - offSet)
-                bodyRectShadow = QtCore.QRect(x + 1, y + offSet, w + 1,
-                                              h - offSet + 1)
-                painter.drawRect(bodyRect)
-
-                pen.setColor(self.palette().color(QtGui.QPalette.Light))
-                pen.setWidthF(0.4)
-                painter.setPen(pen)
-
-                painter.drawRect(bodyRectShadow)
-
-        # draw a boxed style
-        elif self._rolloutStyle == 1:
-            if self.isCollapsed():
-                arect = QtCore.QRect(x + 1, y + 9, w - 1, 4)
-                brect = QtCore.QRect(x, y + 8, w - 1, 4)
-                text = '+'
-            else:
-                arect = QtCore.QRect(x + 1, y + 9, w - 1, h - 9)
-                brect = QtCore.QRect(x, y + 8, w - 1, h - 9)
-                text = '-'
-
-            # draw the borders
-            pen = QtGui.QPen(self.palette().color(QtGui.QPalette.Light))
-            pen.setWidthF(0.6)
-            painter.setPen(pen)
-
-            painter.drawRect(arect)
-
-            pen.setColor(self.palette().color(QtGui.QPalette.Shadow))
-            painter.setPen(pen)
-
-            painter.drawRect(brect)
-
-            painter.setRenderHint(painter.Antialiasing, False)
-            painter.setBrush(
-                self.palette().color(QtGui.QPalette.Window).darker(120))
-            painter.drawRect(x + 10, y + 1, w - 20, 16)
-            painter.drawText(x + 16, y + 1,
-                             w - 32, 16,
-                             QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter,
-                             text)
-            painter.drawText(x + 10, y + 1,
-                             w - 20, 16,
-                             QtCore.Qt.AlignCenter,
-                             self.title())
-
-        if self.dragDropMode():
-            rect = self.dragDropRect()
-
-            # draw the lines
-            l = rect.left()
-            r = rect.right()
-            cy = rect.center().y()
-
-            for y in (cy - 3, cy, cy + 3):
-                painter.drawLine(l, y, r, y)
-
-        painter.end()
-
-    def setCollapsed(self, state=True):
-        if self.isCollapsible():
-            accord = self.accordionWidget()
-            accord.setUpdatesEnabled(False)
-
-            self._collapsed = state
-
-            if state:
-                self.setMinimumHeight(22)
-                self.setMaximumHeight(22)
-                self.widget().setVisible(False)
-            else:
-                self.setMinimumHeight(0)
-                self.setMaximumHeight(1000000)
-                self.widget().setVisible(True)
-
-            self._accordianWidget.emitItemCollapsed(self)
-            accord.setUpdatesEnabled(True)
-
-    def setCollapsible(self, state=True):
-        self._collapsible = state
-
-    def setCustomData(self, key, value):
-        """
-            \remarks	set a custom pointer to information stored on this item
-            \param		key		<str>
-            \param		value	<variant>
-        """
-        self._customData[str(key)] = value
-
-    def setDragDropMode(self, mode):
-        self._dragDropMode = mode
-
-    def setRolloutStyle(self, style):
-        self._rolloutStyle = style
-
-    def showMenu(self):
-        if QtCore.QRect(0, 0, self.width(), 20).contains(
-                self.mapFromGlobal(QtGui.QCursor.pos())):
-            self._accordianWidget.emitItemMenuRequested(self)
-
-    def rolloutStyle(self):
-        return self._rolloutStyle
-
-    def toggleCollapsed(self):
-        # enable signaling here
-        collapse_state = not self.isCollapsed()
-        self.setCollapsed(collapse_state)
-        return collapse_state
-
-    def widget(self):
-        return self._widget
-
-
-class AccordionWidget(QtGui.QScrollArea):
-    itemCollapsed = QtCore.Signal(AccordionItem)
-    itemMenuRequested = QtCore.Signal(AccordionItem)
-    itemDragFailed = QtCore.Signal(AccordionItem)
-    itemsReordered = QtCore.Signal()
-
-    Boxed = 1
-    Rounded = 2
-    Square = 3
-    Maya = 4
-
-    NoDragDrop = 0
-    InternalMove = 1
-
-    def __init__(self, parent):
-
-        QtGui.QScrollArea.__init__(self, parent)
-
-        self.setFrameShape(QtGui.QScrollArea.NoFrame)
-        self.setAutoFillBackground(False)
-        self.setWidgetResizable(True)
-        self.setMouseTracking(True)
-        # self.verticalScrollBar().setMaximumWidth(10)
-
-        widget = QtGui.QWidget(self)
-
-        # define custom properties
-        self._rolloutStyle = AccordionWidget.Rounded
-        self._dragDropMode = AccordionWidget.NoDragDrop
-        self._scrolling = False
-        self._scrollInitY = 0
-        self._scrollInitVal = 0
-        self._itemClass = AccordionItem
-
-        layout = QtGui.QVBoxLayout()
-        layout.setContentsMargins(2, 2, 2, 6)
-        layout.setSpacing(2)
-        layout.addStretch(1)
-
-        widget.setLayout(layout)
-
-        self.setWidget(widget)
-
-    def setSpacing(self, spaceInt):
-        self.widget().layout().setSpacing(spaceInt)
-
-    def addItem(self, title, widget, collapsed=False):
-        self.setUpdatesEnabled(False)
-        item = self._itemClass(self, title, widget)
-        item.setRolloutStyle(self.rolloutStyle())
-        item.setDragDropMode(self.dragDropMode())
-        layout = self.widget().layout()
-        layout.insertWidget(layout.count() - 1, item)
-        layout.setStretchFactor(item, 0)
-
-        if collapsed:
-            item.setCollapsed(collapsed)
-
-        self.setUpdatesEnabled(True)
-        return item
-
-    def clear(self):
-        self.setUpdatesEnabled(False)
-        layout = self.widget().layout()
-        while layout.count() > 1:
-            item = layout.itemAt(0)
-
-            # remove the item from the layout
-            w = item.widget()
-            layout.removeItem(item)
-
-            # close the widget and delete it
-            w.close()
-            w.deleteLater()
-
-        self.setUpdatesEnabled(True)
-
-    def eventFilter(self, object, event):
-        if event.type() == QtCore.QEvent.MouseButtonPress:
-            self.mousePressEvent(event)
-            return True
-
-        elif event.type() == QtCore.QEvent.MouseMove:
-            self.mouseMoveEvent(event)
-            return True
-
-        elif event.type() == QtCore.QEvent.MouseButtonRelease:
-            self.mouseReleaseEvent(event)
-            return True
-
-        return False
-
-    def canScroll(self):
-        return self.verticalScrollBar().maximum() > 0
-
-    def count(self):
-        return self.widget().layout().count() - 1
-
-    def dragDropMode(self):
-        return self._dragDropMode
-
-    def indexOf(self, widget):
-        """
-            \remarks	Searches for widget(not including child layouts).
-                        Returns the index of widget, or -1 if widget is not found
-            \return		<int>
-        """
-        layout = self.widget().layout()
-        for index in range(layout.count()):
-            if layout.itemAt(index).widget().widget() == widget:
-                return index
-        return -1
-
-    def isBoxedMode(self):
-        return self._rolloutStyle == AccordionWidget.Maya
-
-    def itemClass(self):
-        return self._itemClass
-
-    def itemAt(self, index):
-        layout = self.widget().layout()
-        if 0 <= index and index < layout.count() - 1:
-            return layout.itemAt(index).widget()
-        return None
-
-    def emitItemCollapsed(self, item):
-        if not self.signalsBlocked():
-            self.itemCollapsed.emit(item)
-
-    def emitItemDragFailed(self, item):
-        if not self.signalsBlocked():
-            self.itemDragFailed.emit(item)
-
-    def emitItemMenuRequested(self, item):
-        if not self.signalsBlocked():
-            self.itemMenuRequested.emit(item)
-
-    def emitItemsReordered(self):
-        if not self.signalsBlocked():
-            self.itemsReordered.emit()
-
-    def enterEvent(self, event):
-        if self.canScroll():
-            QtGui.QApplication.setOverrideCursor(QtCore.Qt.OpenHandCursor)
-
-    def leaveEvent(self, event):
-        if self.canScroll():
-            QtGui.QApplication.restoreOverrideCursor()
-
-    def mouseMoveEvent(self, event):
-        if self._scrolling:
-            sbar = self.verticalScrollBar()
-            smax = sbar.maximum()
-
-            # calculate the distance moved for the moust point
-            dy = event.globalY() - self._scrollInitY
-
-            # calculate the percentage that is of the scroll bar
-            dval = smax * (dy / float(sbar.height()))
-
-            # calculate the new value
-            sbar.setValue(self._scrollInitVal - dval)
-
-        event.accept()
-
-    def mousePressEvent(self, event):
-        # handle a scroll event
-        if event.button() == QtCore.Qt.LeftButton and self.canScroll():
-            self._scrolling = True
-            self._scrollInitY = event.globalY()
-            self._scrollInitVal = self.verticalScrollBar().value()
-
-            QtGui.QApplication.setOverrideCursor(QtCore.Qt.ClosedHandCursor)
-
-        event.accept()
-
-    def mouseReleaseEvent(self, event):
-        if self._scrolling:
-            QtGui.QApplication.restoreOverrideCursor()
-
-        self._scrolling = False
-        self._scrollInitY = 0
-        self._scrollInitVal = 0
-        event.accept()
-
-    def moveItemDown(self, index):
-        layout = self.widget().layout()
-        if (layout.count() - 1) > (index + 1):
-            widget = layout.takeAt(index).widget()
-            layout.insertWidget(index + 1, widget)
-
-    def moveItemUp(self, index):
-        if index > 0:
-            layout = self.widget().layout()
-            widget = layout.takeAt(index).widget()
-            layout.insertWidget(index - 1, widget)
-
-    def setBoxedMode(self, state):
-        if state:
-            self._rolloutStyle = AccordionWidget.Boxed
-        else:
-            self._rolloutStyle = AccordionWidget.Rounded
-
-    def setDragDropMode(self, dragDropMode):
-        self._dragDropMode = dragDropMode
-
-        for item in self.findChildren(AccordionItem):
-            item.setDragDropMode(self._dragDropMode)
-
-    def setItemClass(self, itemClass):
-        self._itemClass = itemClass
-
-    def setRolloutStyle(self, rolloutStyle):
-        self._rolloutStyle = rolloutStyle
-
-        for item in self.findChildren(AccordionItem):
-            item.setRolloutStyle(self._rolloutStyle)
-
-    def rolloutStyle(self):
-        return self._rolloutStyle
-
-    def takeAt(self, index):
-        self.setUpdatesEnabled(False)
-        layout = self.widget().layout()
-        widget = None
-        if 0 <= index and index < layout.count() - 1:
-            item = layout.itemAt(index)
-            widget = item.widget()
-
-            layout.removeItem(item)
-            widget.close()
-        self.setUpdatesEnabled(True)
-        return widget
-
-    def widgetAt(self, index):
-        item = self.itemAt(index)
-        if item:
-            return item.widget()
-        return None
-
-    pyBoxedMode = QtCore.Property('bool', isBoxedMode, setBoxedMode)
 
 
 class ClickLabel(QtWidgets.QLabel):
@@ -651,8 +36,11 @@ class PreviewWidget(QtWidgets.QWidget):
         QtWidgets.QWidget.__init__(self, parent=parent)
 
         # Add attributes
+        self.initialized = False
         self.options_getter = options_getter
         self.preview = ClickLabel()
+        self.preview.setFixedWidth(1280 / 4)
+        self.preview.setFixedHeight(720 / 4)
 
         # region Build
         self.layout = QtWidgets.QVBoxLayout()
@@ -665,8 +53,6 @@ class PreviewWidget(QtWidgets.QWidget):
 
         # Connect widgets to functions
         self.preview.clicked.connect(self.refresh)
-
-        self.refresh()
 
     def refresh(self):
         """Refresh the playblast preview"""
@@ -683,7 +69,7 @@ class PreviewWidget(QtWidgets.QWidget):
 
             tempdir = tempfile.mkdtemp()
 
-            # override some settings
+            # override settings that are constants for the preview
             options = options.copy()
             options['complete_filename'] = os.path.join(tempdir, "temp.jpg")
             options['width'] = 1280 / 4
@@ -704,15 +90,25 @@ class PreviewWidget(QtWidgets.QWidget):
             self.preview.setPixmap(image)
             os.remove(fname)
 
+    def showEvent(self, event):
+        """Initialize when shown"""
+        if not self.initialized:
+            self.refresh()
+
+        self.initialized = True
+        event.accept()
+
 
 class PresetWidget(QtWidgets.QWidget):
-    """
-    Preset WIdget
+    """Preset Widget
 
     Allows the user to set preferences and create presets to load before 
-    capturing
+    capturing.
 
     """
+
+    preset_loaded = QtCore.Signal(dict)
+    config_opened = QtCore.Signal()
 
     label = "Presets"
 
@@ -721,44 +117,48 @@ class PresetWidget(QtWidgets.QWidget):
 
         self.option_widgets = []
 
-        self._layout = QtWidgets.QHBoxLayout()
-        self._layout.setAlignment(QtCore.Qt.AlignCenter)
-        self._layout.setContentsMargins(0, 0, 0, 0)
-        self.setLayout(self._layout)
+        layout = QtWidgets.QHBoxLayout(self)
+        layout.setAlignment(QtCore.Qt.AlignCenter)
+        layout.setContentsMargins(0, 0, 0, 0)
 
-        self.preset_list = QtWidgets.QComboBox()
-        self.preset_list.setFixedWidth(200)
-        self.preset_list.addItem("*")
+        presets = QtWidgets.QComboBox()
+        presets.setFixedWidth(200)
+        presets.addItem("*")
 
         # Icons
-        icon_path = os.path.dirname(__file__)
+        icon_path = os.path.join(os.path.dirname(__file__), "resources")
         save_icon = os.path.join(icon_path, "save.png")
         load_icon = os.path.join(icon_path, "import.png")
         config_icon = os.path.join(icon_path, "config.png")
-        restore_icon = os.path.join(icon_path, "reset.png")
 
         # Create buttons
-        self.preset_save = QtWidgets.QPushButton()
-        self.preset_save.setIcon(QtWidgets.QIcon(save_icon))
-        self.preset_save.setFixedWidth(30)
+        save = QtWidgets.QPushButton()
+        save.setIcon(QtWidgets.QIcon(save_icon))
+        save.setFixedWidth(30)
 
-        self.preset_load = QtWidgets.QPushButton()
-        self.preset_load.setIcon(QtWidgets.QIcon(load_icon))
-        self.preset_load.setFixedWidth(30)
+        load = QtWidgets.QPushButton()
+        load.setIcon(QtWidgets.QIcon(load_icon))
+        load.setFixedWidth(30)
 
-        self.preset_config = QtWidgets.QPushButton()
-        self.preset_config.setIcon(QtWidgets.QIcon(config_icon))
-        self.preset_config.setFixedWidth(30)
+        config = QtWidgets.QPushButton()
+        config.setIcon(QtWidgets.QIcon(config_icon))
+        config.setFixedWidth(30)
 
-        self.preset_reset = QtWidgets.QPushButton()
-        self.preset_reset.setIcon(QtWidgets.QIcon(restore_icon))
-        self.preset_reset.setFixedWidth(30)
+        layout.addWidget(presets)
+        layout.addWidget(save)
+        layout.addWidget(load)
+        layout.addWidget(config)
 
-        self._layout.addWidget(self.preset_list)
-        self._layout.addWidget(self.preset_save)
-        self._layout.addWidget(self.preset_load)
-        self._layout.addWidget(self.preset_config)
-        self._layout.addWidget(self.preset_reset)
+        self.config = config
+        self.load = load
+        self.save = save
+        self.presets = presets
+
+        # Signals
+        save.clicked.connect(self.save_preset)
+        load.clicked.connect(self.import_preset)
+        config.clicked.connect(self.config_opened)
+        presets.currentIndexChanged.connect(self.load_active_preset)
 
     def import_preset(self):
         """Load preset files to override output values"""
@@ -772,25 +172,29 @@ class PresetWidget(QtWidgets.QWidget):
             return
 
         # create new entry in combobox
-        self.load_presets(filename)
+        self.add_preset(filename)
 
         # read file
         return self.load()
 
-    def load(self):
-        """
-        Load a preset from the already listed preset
+    def load_active_preset(self):
+        """Load the active preset.
         
         :return: collection of preset inputs
         :rtype: dict
         """
-        filename = self.preset_list.currentText()
+        filename = self.presets.currentText()
         if filename == "*":
             return {}
 
-        return lib.load_json(filename)
+        preset = lib.load_json(filename)
 
-    def load_presets(self, filename):
+        # Emit preset load signal
+        self.preset_loaded.emit(preset)
+
+        return preset
+
+    def add_preset(self, filename):
         """
         Add the filename to the preset list and set the index to the filename
         :param filename: the filename of the preset loaded
@@ -798,27 +202,27 @@ class PresetWidget(QtWidgets.QWidget):
         
         :return: None 
         """
-        self.preset_list.blockSignals(True)
+        self.presets.blockSignals(True)
         item_index = 0
-        item_count = self.preset_list.count()
+        item_count = self.presets.count()
         if item_count > 1:
-            current_items = [self.preset_list.itemText(i)
+            current_items = [self.presets.itemText(i)
                              for i in range(item_count)]
 
             # get index of the item from the combobox
             if filename not in current_items:
-                self.preset_list.addItem(filename)
+                self.presets.addItem(filename)
 
-            item_index = self.preset_list.findText(filename)
+            item_index = self.presets.findText(filename)
         else:
-            self.preset_list.addItem(filename)
+            self.presets.addItem(filename)
             item_index += 1
 
         # select item
-        self.preset_list.setCurrentIndex(item_index)
-        self.preset_list.blockSignals(False)
+        self.presets.setCurrentIndex(item_index)
+        self.presets.blockSignals(False)
 
-    def save_presets(self, inputs):
+    def save_preset(self, inputs):
         """Save inputs to a file"""
 
         filters = "Text file (*.json)"
@@ -837,20 +241,22 @@ class PresetWidget(QtWidgets.QWidget):
 
     def get_presets(self):
         """Return all currently listed presets"""
-        configurations = [self.preset_list.itemText(i) for
-                          i in range(self.preset_list.count())]
+        configurations = [self.presets.itemText(i) for
+                          i in range(self.presets.count())]
 
         return configurations
 
     def get_preset(self, filename):
-        if filename not in self.get_presets():
-            self.preset_list.addItem(filename)
-        idx = self.preset_list.findText(filename)
-        self.preset_list.setCurrentIndex(idx)
+
+        assert filename in self.get_presets()
+
+        # if filename not in self.get_presets():
+        #     self.preset_list.addItem(filename)
+        idx = self.presets.findText(filename)
+        self.presets.setCurrentIndex(idx)
 
     def apply_inputs(self, settings):
-        """
-        Apply saved settings of previous session
+        """Apply saved settings of previous session
 
         :param settings: collection of settings based on widget label
         :type settings: dict
@@ -873,7 +279,6 @@ class App(QtWidgets.QWidget):
     """
 
     # Signals
-
     options_changed = QtCore.Signal(dict)
     viewer_start = QtCore.Signal(dict)
 
@@ -886,13 +291,10 @@ class App(QtWidgets.QWidget):
 
         # Settings
         self.settingfile = self._ensure_config_exist()
-        self._translated_settings = {}
-
-        # Add attributes
-        # List of widgets used
-        self.presetwidget = None
-        self.option_widgets = []
-        self.configuration_widgets = []
+        self.plugins = {
+            "app": list(),
+            "config": list()
+        }
 
         # region Set Attributes
         self.setObjectName(objectname)
@@ -921,9 +323,8 @@ class App(QtWidgets.QWidget):
         self.widgetlibrary.addItem("Presets", self.presetwidget)
 
         # add advanced configuration widgets
-        config_widget = self._process_widget()
-        for widget in config_widget:
-            self.add_output_widget(widget)
+        for widget in plugin.discover():
+            self.add_plugin(widget)
 
         self.layout.addWidget(self.widgetlibrary)
 
@@ -939,19 +340,14 @@ class App(QtWidgets.QWidget):
 
         self.apply_inputs(self._read_widget_inputs())
 
-        self.connections()
-
-    def connections(self):
-
         # default actions
         self.apply_button.clicked.connect(self.capture)
         self.close_apply_button.clicked.connect(self.capture_close)
 
-        # apply connections
-        self.presetwidget.preset_config.clicked.connect(self.advanced_configuration)
-        self.presetwidget.preset_load.clicked.connect(self.apply_imported_inputs)
-        self.presetwidget.preset_save.clicked.connect(self.save_inputs)
-        self.presetwidget.preset_list.currentIndexChanged.connect(self.apply_loaded_inputs)
+        # signals and slots
+        self.presetwidget.config_opened.connect(self.advanced_configuration)
+        self.presetwidget.preset_loaded.connect(self.apply_inputs)
+        self.presetwidget.save.clicked.connect(self.save_preset)
 
     def capture(self):
         options = self.get_outputs()
@@ -961,70 +357,24 @@ class App(QtWidgets.QWidget):
         self.capture()
         self.close()
 
-    def _process_widget(self):
-        """
-        Get all the widgets from the widgets module
-        :return: list of filtered widgets
-        :rtype: list
-        """
-
-        plugins = []
-        for name in dir(widgets):
-            # ignore possible build in imported classes
-            if not name[0].isupper():
-                continue
-
-            # skip private stuff
-            if name.startswith("_"):
-                continue
-
-            # check if part of module is a class
-            obj = getattr(widgets, name)
-            if not inspect.isclass(obj):
-                continue
-
-            # ignore the options plugin
-            if obj == widgets.OptionsPlugin:
-                continue
-
-            plugins.append(obj)
-
-        return plugins
-
     def advanced_configuration(self):
         """Show the advanced configuration"""
 
-        self.configuration_dialog = QtWidgets.QDialog(self)
-        self.configuration_dialog.setModal(True)
+        dialog = QtWidgets.QDialog(self)
+        dialog.setModal(True)
 
         config_layout = QtWidgets.QVBoxLayout()
-        for widget in self.configuration_widgets:
+        for widget in self.plugins["config"]:
             groupwidget = QtWidgets.QGroupBox(widget.label)
             group_layout = QtWidgets.QVBoxLayout()
             group_layout.addWidget(widget)
             groupwidget.setLayout(group_layout)
             config_layout.addWidget(groupwidget)
 
-        self.configuration_dialog.setLayout(config_layout)
-        self.configuration_dialog.show()
+        dialog.setLayout(config_layout)
+        dialog.show()
 
-    def apply_imported_inputs(self):
-        """Apply imported preset file"""
-
-        inputs = self.presetwidget.import_preset()
-        self.apply_inputs(inputs)
-
-    def apply_loaded_inputs(self):
-        """
-        Load input settings from choosen file
-        :return: None
-        """
-        inputs = self.presetwidget.load()
-        if not inputs:
-            return
-        self.apply_inputs(inputs)
-
-    def add_output_widget(self, plugin):
+    def add_plugin(self, plugin):
         """Add an options widget plug-in to the UI"""
 
         if plugin.section not in self.application_sections:
@@ -1036,20 +386,21 @@ class App(QtWidgets.QWidget):
         widget.options_changed.connect(self.on_widget_settings_changed)
 
         if widget.section == "config":
-            self.configuration_widgets.append(widget)
+            self.plugins["config"].append(widget)
             return
 
-        if not widget.hidden:
-            item = self.widgetlibrary.addItem(widget.label, widget)
-            widget.label_changed.connect(item.setTitle)
+        elif widget.section == "app":
 
-        # connect change behaviour
-        self.option_widgets.append(widget)
+            if not widget.hidden:
+                item = self.widgetlibrary.addItem(widget.label, widget)
+
+                # connect label change behaviour
+                widget.label_changed.connect(item.setTitle)
+
+            self.plugins["app"].append(widget)
 
     def get_outputs(self):
-        """
-        Collect all options set of all the widgets listed in the in the
-        option_widgets attribute of the main app
+        """Return the settings for a capture as currently set in the Application.
 
         :return: a collection of settings
         :rtype: dict
@@ -1057,23 +408,23 @@ class App(QtWidgets.QWidget):
 
         # Get settings from widgets
         outputs = dict()
-        for widget in self.option_widgets + self.configuration_widgets:
+        for widget in self._get_plugin_widgets():
             outputs.update(widget.get_outputs())
 
         return outputs
 
     def on_widget_settings_changed(self):
         self.options_changed.emit(self.get_outputs)
-        self.presetwidget.preset_list.setCurrentIndex(1)
+        self.presetwidget.presets.setCurrentIndex(1)
 
     # configuration related functions
 
     def _ensure_config_exist(self):
-        """
-        Check if ini file exists in user's folder else create ini file to 
-        write to 
-        :return: filepath of the ini file
+        """Create the configuration file if it does not exist yet.
+        
+        :return: filepath of the configuration file
         :rtype: unicode
+        
         """
 
         userdir = os.path.expanduser("~")
@@ -1092,7 +443,7 @@ class App(QtWidgets.QWidget):
         """Store all used widget settings in the local json file"""
 
         inputs = dict()
-        config_widgets = self.option_widgets + self.configuration_widgets
+        config_widgets = self._get_plugin_widgets()
         for widget in config_widgets:
             settings = widget.get_inputs()
             if not isinstance(settings, dict):
@@ -1103,7 +454,7 @@ class App(QtWidgets.QWidget):
             inputs[widget.id] = settings
 
         with open(self.settingfile, "w") as f:
-            print("FILE : {}".format(f))
+            log.debug("Writing JSON file: {0}".format(f))
             json.dump(inputs, f, sort_keys=True,
                       indent=4, separators=(',', ': '))
 
@@ -1115,25 +466,40 @@ class App(QtWidgets.QWidget):
 
         try:
             with open(self.settingfile, "r") as f:
-                print("FILE : {}".format(f))
+                log.debug("Reading JSON file: {0}".format(f))
                 inputs = json.load(f)
         except ValueError as error:
             log.error(str(error))
 
         return inputs
 
-    def apply_inputs(self, inputs):
+    def _get_plugin_widgets(self):
+        """List all plug-in widgets.
+        
+        :return: The plug-in widgets in *all* sections
+        :rtype: list
+        
         """
-        Apply all the settings of the widgets
+
+        widgets = list()
+        for section in self.plugins.values():
+            widgets.extend(section)
+
+        return widgets
+
+    def apply_inputs(self, inputs):
+        """Apply all the settings of the widgets
         
         :param inputs: collection of input values based on the GUI
         :type inputs: dict
         
         :return: None 
         """
-        option_widgets = self.option_widgets + self.configuration_widgets
+        if not inputs:
+            return
+
         sorted_widgets = dict((widget.id, widget) for
-                              widget in option_widgets)
+                              widget in self._get_plugin_widgets())
 
         # iterate over the sorted widgets to apply the settings
         for widget_id, widget in sorted_widgets.items():
@@ -1142,10 +508,11 @@ class App(QtWidgets.QWidget):
                 continue
             widget.apply_inputs(widget_inputs)
 
-    def save_inputs(self):
-        """Seve the inputs of all the assets"""
+    def save_preset(self):
+        """Save the inputs of all the plugins in a preset"""
+
         inputs = self.get_outputs()
-        filename = self.presetwidget.save_presets(inputs)
+        filename = self.presetwidget.save_preset(inputs)
 
         self.presetwidget.get_preset(filename)
