+++ capture_gui/resources/config.png
