+++ capture_gui/resources/reset.png
