+++ capture_gui/resources/save.png
